pyproject.toml
[tool.poetry]
name = "projet-e4"
version = "0.1.0"
description = ""
authors = ["alphae-nix <18laurent@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Sphinx = "^4.4.0"
sphinx-rtd-theme = "^1.0.0"
myst-parser = "^0.16.1"
python-whois = "^0.7.3"
click = "^8.0.4"
rich = "^11.2.0"
shodan = "^1.26.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
